refactor: 입력값 검증 기능 Validator Class로 분리
