+++ src/main/java/vendingmachine/Change.java
@@ -1,29 +1,38 @@
 package vendingmachine;
 
-import camp.nextstep.edu.missionutils.Console;
-
 import java.util.HashMap;
 
+import camp.nextstep.edu.missionutils.Console;
+
 public class Change {
-    private static final String ERROR_MULTIPLE_OF_TEN = "1원 단위는 허용되지 않습니다.";
-    private static final String ERROR_GREATER_THAN_ZERO = "0 이하의 숫자를 입력할 수 없습니다.";
-    private static final String ERROR_ONLY_INTEGER = "금액은 숫자만 입력 가능합니다.";
+    private static final Validator validator = new Validator();
     private static HashMap<Coin, Integer> coinMap = new HashMap<>();
     private static InputMessage inputMessage = new InputMessage();
+    private static OutputMessage outputMessage = new OutputMessage();
+
+    public int inputInitialTotalChange() {
+        while (true) {
+            try {
+                return insertChange();
+            } catch (IllegalArgumentException exception) {
+                outputMessage.printErrorMessage(exception.getMessage());
+            }
+        }
+    }
 
     public int insertChange() {
         inputMessage.printInsertCoinMessage();
         String changeInVendingMachine = Console.readLine();
-        int change = this.validateOnlyInteger(changeInVendingMachine);
-        isGreatThanZero(change);
-        isMultipleOfTen(change);
+        int change = validator.validateOnlyInteger(changeInVendingMachine);
+        validator.isGreatThanZero(change);
+        validator.isMultipleOfTen(change);
         return change;
     }
 
     public void returnChange(int customerMoney) {
         inputMessage.printChangeMessage();
         for (Coin coin : Coin.values()) {
-            int amount = coinMap.get(coin); // amount
+            int amount = coinMap.get(coin);
             int returnAmount = coin.returnChange(customerMoney, amount);
             coinMap.put(coin, amount - returnAmount);
             customerMoney -= coin.calcChangePrice(returnAmount);
@@ -44,26 +53,6 @@ public class Change {
         inputMessage.printHaveCoinMessage();
         for (Coin coin : Coin.values()) {
             coin.printNumberOfCoin(coinMap.get(coin));
-        }
-    }
-
-    private int validateOnlyInteger(String insertChange) {
-        try {
-            return Integer.parseInt(insertChange);
-        } catch (NumberFormatException exception) {
-            throw new IllegalArgumentException(ERROR_ONLY_INTEGER);
-        }
-    }
-
-    private void isMultipleOfTen(int change) {
-        if (change % 10 != 0) {
-            throw new IllegalArgumentException(ERROR_MULTIPLE_OF_TEN);
-        }
-    }
-
-    private void isGreatThanZero(int change) {
-        if (change <= 0) {
-            throw new IllegalArgumentException(ERROR_GREATER_THAN_ZERO);
         }
     }
 }
